moves lemmas_60k.js to corpuses
